pyproject.toml
[project]
name = "mcp-clients"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-mcp-adapters>=0.2.1",
    "langchain",
    "python-dotenv",
    "streamlit",
    "langchain-openai>=1.1.7",
    "mcp[cli]>=1.26.0",
    "fastmcp>=2.14.5",
]
